Revert "Revert "App structure fix"" (#14)
