+++ package.json
@@ -21,7 +21,7 @@
     "@hookform/resolvers": "^3.3.2",
     "@mui/icons-material": "^5.14.19",
     "@mui/material": "^5.14.20",
-    "@tanstack/react-query": "^5.13.4",
+    "@tanstack/react-query": "^5.14.1",
     "firebase": "^10.7.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
