+++ src/components/Footer/constants.ts
@@ -1,0 +1,5 @@
+export const ProfileLinks = [
+  "https://github.com/ffriday",
+  "https://github.com/krkate",
+  "https://github.com/romankadevich",
+];
