+++ src/layouts/RootLayout.tsx
@@ -1,6 +1,6 @@
-import { Header } from "../components/header/header";
-import { Footer } from "../components/footer/footer";
 import { Outlet } from "react-router-dom";
+import { Footer } from "../components/Footer/Footer";
+import { Header } from "../components/Header/Header";
 
 export const RootLayout = () => {
   return (
