+++ src/pages/SignUpPage/sign-up.module.scss
